+++ .github/scripts/scrapejs/saveJson.js
@@ -1,10 +1,14 @@
-const { JSON_INDENT_SIZE } = require('./variables');
 const fs = require('fs');
+
+const WriteFileConfig = {
+  JSON_INDENT_SIZE: 2,
+  CHARSET: 'utf-8'
+};
 
 const saveJson = (data, paths) => {
   for (const path of paths) {
     console.log('Сохраняю данные по пути: ', path);
-    fs.writeFileSync(path, JSON.stringify(data, null, JSON_INDENT_SIZE), 'utf-8');
+    fs.writeFileSync(path, JSON.stringify(data, null, WriteFileConfig.JSON_INDENT_SIZE), WriteFileConfig.CHARSET);
   }
 };
